fix(Contacts): OZF-492 Loosening rules for required Contact Types
Required Contact types are no longer strictly validated on all updates
to non-draft listings.  Instead, the required contact types are fully
validated when the listing is submitted (that is, when it is
tranistioned into the PENDING approvalStatus).  After that, the listing
is only checked to make sure that a required contact type was not
removed from it.  Required Contact Types that are added to the system
after the listing is submitted are not strictly required.  As before,
required contact types are not validated at all on draft listings